chore: rename Components to components
