+++ src/components/Autofill/AutoFillInfoStrip/index.js
