+++ src/components/Autofill/AutofillSaleCategories/index.js
