+++ src/components/Categories/CategoryBlock.js
