+++ src/components/Categories/CategoryTemplate.js
