+++ src/components/CopyButton/index.js
