+++ src/components/Hero/HeroTemplate.js
