+++ src/components/InfoStrip/InfoStripContainer.js
