+++ src/components/Sale/SaleCategoriesTemplate.js
