+++ src/components/SaleCategories/SaleCategoryBlock.js
