+++ src/components/SaleCategories/SaleCategoryTemplate.js
