+++ src/components/Ticker/TickerBlock.js
