+++ src/components/Ticker/TickerTemplate.js
