+++ src/components/styles/ContentBlocks.js
