+++ src/context/updateHtml.js
@@ -1,14 +1,14 @@
-import InfoStripTemplate from '../Components/InfoStrip/InfoStripTemplate'
-import HeroTemplate from '../Components/Hero/HeroTemplate'
-import LowerTemplate from '../Components/Lower/LowerTemplate'
-import SaleTemplate from '../Components/Sale/SaleTemplate'
-import SaleCategoriesTemplate from '../Components/SaleCategories/SaleCategoryTemplate'
-import MegaBannerWidget from '../Components/MegaBanner'
-import TickerTemplate from '../Components/Ticker/TickerTemplate'
-import CategoryTemplate from '../Components/Categories/CategoryTemplate'
+import InfoStripTemplate from '../components/InfoStrip/InfoStripTemplate'
+import HeroTemplate from '../components/Hero/HeroTemplate'
+import LowerTemplate from '../components/Lower/LowerTemplate'
+import SaleTemplate from '../components/Sale/SaleTemplate'
+import SaleCategoriesTemplate from '../components/SaleCategories/SaleCategoryTemplate'
+import MegaBannerWidget from '../components/MegaBanner'
+import TickerTemplate from '../components/Ticker/TickerTemplate'
+import CategoryTemplate from '../components/Categories/CategoryTemplate'
 import addPlaceholderImageCSS from '../lib/addPlaceholderImageCSS'
-import { AppDownloadTemplate } from '../Components/AppDownload'
-import { SaleTextBannerTemplate } from '../Components/SaleTextBanner'
+import { AppDownloadTemplate } from '../components/AppDownload'
+import { SaleTextBannerTemplate } from '../components/SaleTextBanner'
 
 const hasContent = (state) => {
   if (Array.isArray(state)) {
